requirements.txt
Django==4.2.16
djangorestframework==3.15.2
rest_framework_simplejwt==5.3.1
mysqlclient==2.2.4
PyMySQL==1.1.1
django-redis==5.4.0


